+++ src/platform/RobloxPlatform.cpp
@@ -681,11 +681,17 @@ std::optional<Luau::AutocompleteEntryMap> RobloxPlatform::completionCallback(
             auto ctv = ctx.value();
             while (ctv)
             {
-                for (auto& [propName, _] : ctv->props)
+                for (auto& [propName, prop] : ctv->props)
                 {
-                    result.insert_or_assign(
-                        propName, Luau::AutocompleteEntry{Luau::AutocompleteEntryKind::String, workspaceFolder->frontend.builtinTypes->stringType,
-                                      false, false, Luau::TypeCorrectKind::Correct});
+                    // Don't include functions or events
+                    auto ty = Luau::follow(prop.type());
+                    if (Luau::get<Luau::FunctionType>(ty) || Luau::isOverloadedFunction(ty))
+                        continue;
+                    else if (auto ttv = Luau::get<Luau::TableType>(ty); ttv && ttv->name && ttv->name.value() == "RBXScriptSignal")
+                        continue;
+
+                    result.insert_or_assign(propName, Luau::AutocompleteEntry{Luau::AutocompleteEntryKind::String, frontend.builtinTypes->stringType,
+                                                          false, false, Luau::TypeCorrectKind::Correct});
                 }
                 if (ctv->parent)
                     ctv = Luau::get<Luau::ClassType>(*ctv->parent);